separate InstanceOverview into table and panel, make sure sourceApplicationForm overflows with scroll
