+++ src/features/instanceOverview/components/InstancePanel.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+
+const InstancePanel: React.FunctionComponent<any> = (props) => {
+
+    return (
+        <div>
+        </div>
+    );
+}
+
+export default InstancePanel;
